Bug 1616931 - send log messages via worker-runner where supported
With `--with-worker-runner`:
```
(go1.13.7) (v12.16.1) (sandbox) dustin@lamport ~/p/taskcluster/workers/generic-worker [bug1616931*] $  ./generic-worker-simple-linux-amd64 run --with-worker-runner
~{"type": "welcome", "capabilities": ["log"]} // I typed this
~{"capabilities":["log"],"type":"hello"}
~{"textPayload":"Loading generic-worker config file '/home/dustin/p/taskcluster/workers/generic-worker/generic-worker.config'...","type":"log"}
..
```

Without, it still works like it always did:
```
(go1.13.7) (v12.16.1) (sandbox) dustin@lamport ~/p/taskcluster/workers/generic-worker [bug1616931*] $  ./generic-worker-simple-linux-amd64 run
2020/02/28 22:55:04Z Loading generic-worker config file '/home/dustin/p/taskcluster/workers/generic-worker/generic-worker.config'...
..
```
